apps/server/requirements.txt
asgiref==3.8.1
black==25.1.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
